backend/requirements.txt
requests
pandas
matplotlib
PyJWT
python-dotenv
cryptography
pymongo
alpaca-py
tqdm
